requirements.txt
PyYAML==6.0.2
watchdog==4.0.1
